+++ packages/realm-web-integration-tests/harness/import-realm-app.ts
@@ -26,17 +26,29 @@ const MDB_REALM_USERNAME =
     process.env.MDB_REALM_USERNAME || "unique_user@domain.com";
 const MDB_REALM_PASSWORD = process.env.MDB_REALM_PASSWORD || "password";
 
+const MDB_REALM_APP_ID = process.env.MDB_REALM_APP_ID;
+
 export async function importRealmApp() {
     // Create a new MongoDBRealmService
     const baseUrl = MDB_REALM_BASE_URL;
-    const importer = new RealmAppImporter({
-        baseUrl,
-        username: MDB_REALM_USERNAME,
-        password: MDB_REALM_PASSWORD,
-        appsDirectoryPath: path.resolve(__dirname, "../imported-apps"),
-        stitchConfigPath: path.resolve(__dirname, "../stitch-config.json"),
-    });
-    const appTemplatePath = path.resolve(__dirname, "../my-test-app-template");
-    const { appId } = await importer.importApp(appTemplatePath);
-    return { appId, baseUrl };
+    if (MDB_REALM_APP_ID) {
+        console.log(
+            `Skipping import of the app (MDB_REALM_APP_ID = ${MDB_REALM_APP_ID})`,
+        );
+        return { appId: MDB_REALM_APP_ID, baseUrl };
+    } else {
+        const importer = new RealmAppImporter({
+            baseUrl,
+            username: MDB_REALM_USERNAME,
+            password: MDB_REALM_PASSWORD,
+            appsDirectoryPath: path.resolve(__dirname, "../imported-apps"),
+            stitchConfigPath: path.resolve(__dirname, "../stitch-config.json"),
+        });
+        const appTemplatePath = path.resolve(
+            __dirname,
+            "../my-test-app-template",
+        );
+        const { appId } = await importer.importApp(appTemplatePath);
+        return { appId, baseUrl };
+    }
 }