+++ packages/realm-web/src/Credentials.ts
@@ -16,9 +16,15 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-type AnonymousPayload = Realm.Credentials.AnonymousPayload;
-type ApiKeyPayload = Realm.Credentials.ApiKeyPayload;
-type EmailPasswordPayload = Realm.Credentials.EmailPasswordPayload;
+export type AnonymousPayload = Realm.Credentials.AnonymousPayload;
+export type ApiKeyPayload = Realm.Credentials.ApiKeyPayload;
+export type EmailPasswordPayload = Realm.Credentials.EmailPasswordPayload;
+export type OAuth2RedirectPayload = Realm.Credentials.OAuth2RedirectPayload;
+export type GooglePayload = Realm.Credentials.GooglePayload;
+export type FacebookPayload = Realm.Credentials.FacebookPayload;
+export type FunctionPayload = Realm.Credentials.FunctionPayload;
+export type JWTPayload = Realm.Credentials.JWTPayload;
+export type ApplePayload = Realm.Credentials.ApplePayload;
 
 /**
  * Instances of this class can be passed to the `app.logIn` method to authenticate an end-user.
@@ -26,7 +32,7 @@ type EmailPasswordPayload = Realm.Credentials.EmailPasswordPayload;
 export class Credentials<PayloadType extends object>
     implements Realm.Credentials<PayloadType> {
     /**
-     * Creates credentials that logs in using the [Anonymous Provider](https://docs.mongodb.com/stitch/authentication/anonymous/).
+     * Creates credentials that logs in using the [Anonymous Provider](https://docs.mongodb.com/realm/authentication/anonymous/).
      *
      * @returns The credentials instance, which can be passed to `app.logIn`.
      */
@@ -35,7 +41,7 @@ export class Credentials<PayloadType extends object>
     }
 
     /**
-     * Creates credentials that logs in using the [API Key Provider](https://docs.mongodb.com/stitch/authentication/api-key/).
+     * Creates credentials that logs in using the [API Key Provider](https://docs.mongodb.com/realm/authentication/api-key/).
      *
      * @param key The secret content of the API key.
      * @returns The credentials instance, which can be passed to `app.logIn`.
@@ -47,7 +53,7 @@ export class Credentials<PayloadType extends object>
     }
 
     /**
-     * Creates credentials that logs in using the [Email/Password Provider](https://docs.mongodb.com/stitch/authentication/userpass/).
+     * Creates credentials that logs in using the [Email/Password Provider](https://docs.mongodb.com/realm/authentication/email-password/).
      * Note: This was formerly known as the "Username/Password" provider.
      *
      * @param email The end-users email address.
@@ -62,6 +68,91 @@ export class Credentials<PayloadType extends object>
                 username: email,
                 password,
             },
+        );
+    }
+
+    /**
+     * Creates credentials that logs in using the [Custom Function Provider](https://docs.mongodb.com/realm/authentication/custom-function/).
+     *
+     * @param payload The custom payload as expected by the server.
+     * @returns The credentials instance, which can be passed to `app.logIn`.
+     */
+    static function<PayloadType extends FunctionPayload = FunctionPayload>(
+        payload: PayloadType,
+    ) {
+        return new Credentials<PayloadType>(
+            "custom-function",
+            "custom-function",
+            payload,
+        );
+    }
+
+    /**
+     * Creates credentials that logs in using the [Custom JWT Provider](https://docs.mongodb.com/realm/authentication/custom-jwt/).
+     *
+     * @param token The JSON Web Token (JWT).
+     * @returns The credentials instance, which can be passed to `app.logIn`.
+     */
+    static jwt(token: string) {
+        return new Credentials<ApiKeyPayload>("custom-token", "custom-token", {
+            token,
+        });
+    }
+
+    /**
+     * Creates credentials that logs in using the [Google Provider](https://docs.mongodb.com/realm/authentication/google/).
+     *
+     * @param redirectUrlOrAuthCode The URL that users should be redirected to or the auth code returned from Google.
+     * @returns The credentials instance, which can be passed to `app.logIn`.
+     */
+    static google<
+        PayloadType extends object = OAuth2RedirectPayload | GooglePayload
+    >(redirectUrlOrAuthCode: string) {
+        return new Credentials<PayloadType>(
+            "oauth2-google",
+            "oauth2-google",
+            redirectUrlOrAuthCode.indexOf("://") !== -1
+                ? { redirectUrl: redirectUrlOrAuthCode }
+                : { authCode: redirectUrlOrAuthCode },
+        );
+    }
+
+    /**
+     * Creates credentials that logs in using the [Facebook Provider](https://docs.mongodb.com/realm/authentication/facebook/).
+     *
+     * @param redirectUrlOrAccessToken The URL that users should be redirected to or the auth code returned from Facebook.
+     * @returns The credentials instance, which can be passed to `app.logIn`.
+     */
+    static facebook<
+        PayloadType extends object = OAuth2RedirectPayload | FacebookPayload
+    >(redirectUrlOrAccessToken: string) {
+        return new Credentials<PayloadType>(
+            "oauth2-facebook",
+            "oauth2-facebook",
+            redirectUrlOrAccessToken.indexOf("://") !== -1
+                ? { redirectUrl: redirectUrlOrAccessToken }
+                : { accessToken: redirectUrlOrAccessToken },
+        );
+    }
+
+    /**
+     * Creates credentials that logs in using the [Apple ID Provider](https://docs.mongodb.com/realm/authentication/apple/).
+     *
+     * @param redirectUrlOrIdToken The URL that users should be redirected to or the id_token returned from Apple.
+     * @returns The credentials instance, which can be passed to `app.logIn`.
+     */
+    static apple<
+        PayloadType extends object = OAuth2RedirectPayload | ApplePayload
+    >(redirectUrlOrIdToken: string) {
+        return new Credentials<PayloadType>(
+            "oauth2-apple",
+            "oauth2-apple",
+            redirectUrlOrIdToken.indexOf("://") !== -1
+                ? { redirectUrl: redirectUrlOrIdToken }
+                : {
+                      // eslint-disable-next-line @typescript-eslint/camelcase
+                      id_token: redirectUrlOrIdToken,
+                  },
         );
     }
 
@@ -87,6 +178,27 @@ export class Credentials<PayloadType extends object>
         name: string,
         type: "local-userpass",
         payload: EmailPasswordPayload,
+    );
+    constructor(
+        name: string,
+        type: "custom-function",
+        payload: FunctionPayload,
+    );
+    constructor(name: string, type: "custom-token", payload: JWTPayload);
+    constructor(
+        name: string,
+        type: "oauth2-google",
+        payload: OAuth2RedirectPayload | GooglePayload,
+    );
+    constructor(
+        name: string,
+        type: "oauth2-facebook",
+        payload: OAuth2RedirectPayload | FacebookPayload,
+    );
+    constructor(
+        name: string,
+        type: "oauth2-apple",
+        payload: OAuth2RedirectPayload | ApplePayload,
     );
 
     /**