+++ packages/realm-web/src/FunctionsFactory.ts
@@ -28,11 +28,17 @@ const RESERVED_NAMES = ["inspect", "callFunction"];
  * The body of the request sent to call a remote function.
  */
 interface CallFunctionBody {
-    /** Name of the function. */
+    /**
+     * Name of the function.
+     */
     name: string;
-    /** An array of arguments to pass to the function. */
+    /**
+     * An array of arguments to pass to the function.
+     */
     arguments: any[];
-    /** An optional name of the service in which the function is defined. */
+    /**
+     * An optional name of the service in which the function is defined.
+     */
     service?: string;
 }
 
@@ -40,11 +46,17 @@ interface CallFunctionBody {
  * Pass an object implementing this interface when constructing a functions factory.
  */
 export interface FunctionsFactoryConfiguration {
-    /** An optional name of the service in which functions are defined. */
+    /**
+     * An optional name of the service in which functions are defined.
+     */
     serviceName?: string;
-    /** Call this function to transform the arguments before they're sent to the service. */
+    /**
+     * Call this function to transform the arguments before they're sent to the service.
+     */
     argsTransformation?: (args: any[]) => any[];
-    /** Call this function to transform a response before it's returned to the caller. */
+    /**
+     * Call this function to transform a response before it's returned to the caller.
+     */
     responseTransformation?: (response: any) => any;
 }
 
@@ -68,7 +80,7 @@ function cleanArgs(args: any[]) {
 }
 
 /**
- * Remove keys for any undefined values and serialize to EJSON
+ * Remove keys for any undefined values and serialize to EJSON.
  *
  * @param args The arguments to clean and serialize.
  * @returns The cleaned and serialized arguments.
@@ -79,7 +91,7 @@ function cleanArgsAndSerialize(args: any[]) {
 }
 
 /**
- * Defines how functions are called
+ * Defines how functions are called.
  */
 export class FunctionsFactory {
     /**
@@ -113,21 +125,31 @@ export class FunctionsFactory {
         });
     }
 
-    /** The underlying transport to use when requesting */
+    /**
+     * The underlying transport to use when requesting.
+     */
     private readonly transport: Transport;
 
-    /** An optional name of the service in which functions are defined */
+    /**
+     * An optional name of the service in which functions are defined.
+     */
     private readonly serviceName?: string;
-    /** Call this function to transform the arguments before they're sent to the service */
+
+    /**
+     * Call this function to transform the arguments before they're sent to the service.
+     */
     private readonly argsTransformation?: (args: any[]) => any[];
-    /** Call this function to transform a response before it's returned to the caller */
+
+    /**
+     * Call this function to transform a response before it's returned to the caller.
+     */
     private readonly responseTransformation?: (response: any) => any;
 
     /**
-     * Construct a functions factory
+     * Construct a functions factory.
      *
-     * @param transport The underlying transport to use when sending requests
-     * @param config Additional configuration parameters
+     * @param transport The underlying transport to use when sending requests.
+     * @param config Additional configuration parameters.
      */
     constructor(
         transport: Transport,
@@ -142,10 +164,10 @@ export class FunctionsFactory {
     }
 
     /**
-     * Call a remote function by it's name
+     * Call a remote function by it's name.
      *
-     * @param name Name of the remote function
-     * @param args Arguments to pass to the remote function
+     * @param name Name of the remote function.
+     * @param args Arguments to pass to the remote function.
      * @returns A promise of the value returned when executing the remote function.
      */
     async callFunction(name: string, ...args: any[]): Promise<any> {