+++ packages/realm-web/src/index.ts
@@ -17,13 +17,15 @@
 ////////////////////////////////////////////////////////////////////////////
 
 import { App } from "./App";
+import { OAuth2Helper } from "./OAuth2Helper";
+import { createDefaultStorage } from "./storage";
 
 const appCache: { [id: string]: Realm.App } = {};
 
 /**
  * Get or create a singleton Realm App from an id.
  *
- * @param id The Realm App id visible from the MongoDB Realm UI or a configuration
+ * @param id The Realm App id visible from the MongoDB Realm UI or a configuration.
  * @returns The Realm App instance. Calling this function multiple times with the same id will return the same instance.
  */
 export function app(id: string) {
@@ -33,6 +35,25 @@ export function app(id: string) {
         const instance = new App(id);
         appCache[id] = instance;
         return instance;
+    }
+}
+
+/**
+ * Handle an OAuth 2.0 redirect.
+ *
+ * @param location An optional location to use (defaults to the windows current location).
+ * @param storage Optional storage used to save any results from the location.
+ */
+export function handleAuthRedirect(
+    location = window.location,
+    storage = createDefaultStorage(),
+) {
+    try {
+        const queryString = location.hash.substr(1); // Strip the initial # from the hash
+        OAuth2Helper.handleRedirect(queryString, storage);
+    } catch (err) {
+        // Ensure calling this never throws: It should not interrupt a users flow.
+        console.warn(err);
     }
 }