+++ packages/realm-web/src/storage/PrefixedStorage.ts
@@ -16,14 +16,14 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-import { Storage } from "./Storage";
+import { Storage, StorageChangeListner } from "./Storage";
 
 /**
  * A `Storage` which will prefix a key part to every operation.
  */
 export class PrefixedStorage implements Storage {
     /**
-     * The string separating two parts
+     * The string separating two parts.
      */
     private static PART_SEPARATOR = ":";
 
@@ -80,5 +80,15 @@ export class PrefixedStorage implements Storage {
         return this.storage.clear(
             this.keyPart + PrefixedStorage.PART_SEPARATOR + prefix,
         );
+    }
+
+    /** @inheritdoc */
+    public addListener(listener: StorageChangeListner) {
+        return this.storage.addListener(listener);
+    }
+
+    /** @inheritdoc */
+    public removeListener(listener: StorageChangeListner) {
+        return this.storage.addListener(listener);
     }
 }